requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python
Pillow
numpy
python-jose[cryptography]
passlib[bcrypt]
python-dotenv